+++ frontend/components/Form/useForm.tsx
@@ -5,6 +5,7 @@ import { TextFieldProps, SelectProps } from "@material-ui/core";
 import useTextField from "./useTextField";
 import useSelect from "./useSelect";
 import usePasswordField from "./usePasswordField";
+import useDateRangeField from "./useDateRange";
 
 interface useETKFormSchema {
   [key: string]: {
@@ -24,6 +25,7 @@ export default function useETKForm(props: useETKFormProps) {
   const textfields = {};
   const passwordfields = {};
   const selects = {};
+  const dateranges = {};
 
   for (const field in props.schema) {
     shape[field] = props.schema[field].schema;
@@ -37,6 +39,9 @@ export default function useETKForm(props: useETKFormProps) {
         break;
       case "select":
         selects[field] = props.schema[field].component;
+        break;
+      case "daterange":
+        dateranges[field] = props.schema[field].component;
         break;
     }
   }
@@ -53,7 +58,9 @@ export default function useETKForm(props: useETKFormProps) {
   const a = useTextField({ fields: textfields, ...form });
   const b = useSelect({ fields: selects, ...form });
   const c = usePasswordField({fields: passwordfields, ...form});
-  const fields = Object.assign({}, a, b, c);
+  const d = useDateRangeField({fields: dateranges, ...form});
+  
+  const fields = Object.assign({}, a, b, c, d);
 
   return {
     fields,