+++ frontend/components/Panel.tsx
@@ -29,6 +29,7 @@ const ETKPanelPanels = {
   start: dynamic(() => import("./Panel/Start/Index")),
   import: dynamic(() => import("./Import/Panel/Index")),
   newTree: dynamic(() => import("./Tree/Form")),
+  newIntervention: dynamic(() => import("./Interventions/Form"))
 };
 
 const ETKPanel: React.FC<ETKPanelProps> = (props) => {